requirements.txt
python-telegram-bot==20.8
web3==6.10.0
pymongo==4.7.1
requests==2.31.0
